Apply automatically generated source from `agora-gen` - extern (C++)
